Create widgets folder
