+++ lib/routes.dart
@@ -18,7 +18,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:picos/screens/forgot_password_screen/forgot_password_screen.dart';
 import 'package:picos/screens/study_nurse_screen/follow_up_screen/edit_follow_up_screen.dart';
-import 'package:picos/screens/study_nurse_screen/follow_up_screen/follow_up_list.dart';
+import 'package:picos/screens/study_nurse_screen/follow_up_screen/widgets/follow_up_list.dart';
 import 'package:picos/screens/legals_screen.dart';
 import 'package:picos/screens/login_screen.dart';
 import 'package:picos/screens/my_documents_screen/add_document_screen.dart';
