+++ lib/screens/my_medications_screen/widgets/medication_card.dart
@@ -19,11 +19,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:picos/models/medication.dart';
-import 'package:picos/screens/my_medications_screen/medication_card_tile.dart';
+import 'package:picos/screens/my_medications_screen/widgets/medication_card_tile.dart';
 import 'package:picos/widgets/picos_list_card.dart';
 
-import '../../api/backend_medications_api.dart';
-import '../../state/objects_list_bloc.dart';
+import '../../../api/backend_medications_api.dart';
+import '../../../state/objects_list_bloc.dart';
 
 /// The card displaying a medication plan.
 class MedicationCard extends StatelessWidget {
