+++ lib/screens/my_medications_screen/widgets/medications_list.dart
@@ -17,26 +17,26 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:picos/models/abstract_database_object.dart';
-import '../../api/backend_therapies_api.dart';
-import '../../models/therapy.dart';
+import '../../../api/backend_medications_api.dart';
+import '../../../models/medication.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import '../../state/objects_list_bloc.dart';
-import 'therapy_item.dart';
+import '../../../state/objects_list_bloc.dart';
+import 'medication_card.dart';
 
-/// A List with all therapies.
-class TherapiesList extends StatefulWidget {
-  /// Creates TherapiesList.
-  const TherapiesList({Key? key}) : super(key: key);
+/// A List with all medications.
+class MedicationsList extends StatefulWidget {
+  /// Creates the medication list.
+  const MedicationsList({Key? key}) : super(key: key);
 
   @override
-  State<TherapiesList> createState() => _TherapiesListState();
+  State<MedicationsList> createState() => _MedicationsListState();
 }
 
-class _TherapiesListState extends State<TherapiesList> {
+class _MedicationsListState extends State<MedicationsList> {
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<ObjectsListBloc<BackendTherapiesApi>, ObjectsListState>(
+    return BlocBuilder<ObjectsListBloc<BackendMedicationsApi>,
+        ObjectsListState>(
       builder: (BuildContext context, ObjectsListState state) {
         if (state.status == ObjectsListStatus.initial ||
             state.status == ObjectsListStatus.loading) {
@@ -51,24 +51,10 @@ class _TherapiesListState extends State<TherapiesList> {
           );
         }
 
-        state.objectsList.sort(
-          (AbstractDatabaseObject a, AbstractDatabaseObject b) =>
-              (b as Therapy).date.compareTo((a as Therapy).date),
-        );
-
-        return ListView.separated(
+        return ListView.builder(
           itemCount: state.objectsList.length,
           itemBuilder: (BuildContext context, int index) {
-            return TherapyItem(state.objectsList[index] as Therapy);
-          },
-          separatorBuilder: (BuildContext context, int index) {
-            return const Padding(
-              padding: EdgeInsets.symmetric(horizontal: 15),
-              child: Divider(
-                thickness: 1,
-                height: 0,
-              ),
-            );
+            return MedicationCard(state.objectsList[index] as Medication);
           },
         );
       },
