+++ lib/screens/my_visits_screen/widgets/visit_item.dart
